Split up CommonFunctionsTests
